pyproject.toml
[tool.poetry]
name = "librarian"
version = "0.1.0"
description = ""
authors = ["Nathan Evans <nathan@de1ux.com>"]

[tool.poetry.dependencies]
python = "^3.9"
google-cloud-storage = "^1.34.0"
google-cloud-vision = "^2.0.0"
psycopg2-binary = "^2.8.6"
imagehash = "^4.2.0"
Django = "^3.2.2"
django-rest-framework = "^0.1.0"
django-cors-headers = "^3.7.0"
libnfs = "1.0.post4"
python-env-utils = "^0.4.1"

[tool.poetry.dev-dependencies]
black = "^21.5b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
